# ENH: Optimize state probability expression parsing