@nawtrey 2024-08-03
I ran the KDA manuscript code (see [this comment](https://github.com/Becksteinlab/kda/pull/88#issuecomment-2267022303)) on this branch and it completed without issue. I'll go ahead and merge this so we can start taking advantage of the performance benefits.